pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Valery <teslenko.70@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.2.2"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.3.3"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
